- Added UI for transactions cells - Added logic for counting the total amount spent

+++ PersonalFinances/Model/Transaction.swift
@@ -9,5 +9,5 @@ import Foundation
 
 struct Transaction {
     let transactionName: String
-    let transactionAmount: Float
+    let transactionAmount: Double
 }
